crud boards working

+++ src/main/java/com/uiowa/fse_project/model/UserDtls.java
@@ -13,9 +13,7 @@ public class UserDtls {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-
-	//private String username;
+	private long id;
 
 	private String firstName;
